requirements.txt
asyncpg==0.30.0
boto3==1.35.57
botocore==1.35.57
fastapi==0.115.4
fitz==0.0.1.dev2
groq==0.11.0
pydantic==2.9.2
python-dotenv==1.0.1
